Fixes in _CGUI while I was familiarizing myself with the code.
Fixed the following errors:
In the command-line interface, when inputting desired languages for
splitting, it would get off count if incorrect codes were input, then
keep requesting codes indefinitely.  Run the old code, select option 1
or 3, say you want more than one language, and when prompted for
language codes write invalid codes.  The count goes up and up and never
stops (unless -quit is entered, of course, but that just quits.)
Also, when asking for the number of desired languages in options 1 and
3, it accepted any input containing no alphabetical characters, even if
it was not a valid number (like the empty string, or numbers and
spaces.)  Now it is fixed.